refactor cmm to lib
