+++ src/lib.rs
@@ -16,6 +16,15 @@ pub mod settings;
 pub mod tracer;
 pub mod tracing_layers;
 
+use axum::{
+    Json,
+    response::{Html, IntoResponse},
+    http::{StatusCode, Uri, header::{self, HeaderMap, HeaderName}},
+};
+use std::path::PathBuf;
+use std::io::{BufWriter, Cursor};
+use image::ImageFormat;
+
 /// Test utilities.
 #[cfg(any(test, feature = "test_utils"))]
 #[cfg_attr(docsrs, doc(cfg(feature = "test_utils")))]
@@ -24,3 +33,21 @@ pub mod test_utils;
 pub fn add(a: i32, b: i32) -> i32 {
     a + b
 }
+// stream.
+pub async fn get_image() -> impl axum::response::IntoResponse {
+    let img_path = PathBuf::from("assets/").join("a_logo.png");
+    let image = image::io::Reader::open(&img_path).unwrap().decode().unwrap();
+    let mut buffer = BufWriter::new(Cursor::new(Vec::new()));
+    image.write_to(&mut buffer, ImageFormat::Png).unwrap();
+    let bytes: Vec<u8> = buffer.into_inner().unwrap().into_inner();
+    (
+        axum::response::AppendHeaders([(header::CONTENT_TYPE, "image/png")]),
+        bytes
+        // image.into_bytes()
+    )
+}
+
+pub async fn html() -> Html<&'static str> {
+    Html("<p>Hello, World!</p>")
+}
+// roadmap - show webcam on laptop screen.
